3.0.0
I ran "vsce publish major" but that failed because my access token had
expired. The commit with the  version bump was still made by vsce though
so when I ran the command again I was suprised it's sayig 3.0.0
Seems there is no way to unpublish versions on the marketplace yet, so
we have to live with this.